requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
evidently
dill
matplotlib
-e .
